@@ -1,36 +1,34 @@
 {
-	"name": "pinger",
+	"name": "my-app",
+	"private": true,
 	"version": "0.0.1",
+	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
+		"prepare": "svelte-kit sync || echo ''",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^2.0.0",
-		"@sveltejs/adapter-node": "^1.3.0",
-		"@sveltejs/kit": "^1.20.4",
-		"@types/better-sqlite3": "^7.6.9",
-		"@types/cookie": "^0.5.1",
+		"@sveltejs/adapter-node": "^5.4.0",
+		"@sveltejs/kit": "^2.49.2",
+		"@sveltejs/vite-plugin-svelte": "^6.2.1",
+		"@tailwindcss/vite": "^4.1.17",
 		"@types/signale": "^1.4.7",
-		"autoprefixer": "^10.4.17",
-		"drizzle-kit": "^0.20.14",
-		"postcss": "^8.4.35",
-		"svelte": "^4.0.0",
-		"svelte-check": "^3.4.3",
-		"tailwindcss": "^3.4.1",
-		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^4.3.6"
+		"dotenv": "^17.2.3",
+		"prisma": "^7.1.0",
+		"svelte": "^5.43.8",
+		"svelte-check": "^4.3.4",
+		"tailwindcss": "^4.1.17",
+		"typescript": "^5.9.3",
+		"vite": "^7.2.2"
 	},
-	"type": "module",
 	"dependencies": {
-		"@tailwindcss/forms": "^0.5.7",
-		"better-sqlite3": "^9.4.1",
-		"drizzle-orm": "^0.29.3",
-		"nanoid": "^5.0.5",
-		"signale": "^1.4.0"
+		"@prisma/adapter-better-sqlite3": "^7.1.0",
+		"@prisma/client": "^7.1.0",
+		"signale": "^1.4.0",
+		"zod": "^4.1.13"
 	}
 }
