@@ -1,21 +1,21 @@
 <script lang="ts">
+	import { onMount } from 'svelte';
 	import { browser } from '$app/environment';
 	import { invalidateAll } from '$app/navigation';
-	import { onMount } from 'svelte';
+	import { REQUEST_TIMEOUT } from '$lib/config';
 
-	export let interval: number;
-	export let lastFetchDate: number;
+	let { interval, lastPingDate }: { interval: number; lastPingDate: Date } = $props();
 
-	let percent: number = getPercent();
-	$: makeSmaller = percent > 95;
+	let percent: number = $state(getPercent());
+	let currentlyPinging = $derived(percent > 95);
 
 	function getPercent() {
-		const timeElapsed = Date.now() - lastFetchDate;
+		const timeElapsed = Date.now() - lastPingDate.getTime() - REQUEST_TIMEOUT;
 		const newPercent = Math.floor((timeElapsed / interval) * 100);
 
 		if (newPercent > 100) {
 			if (browser) invalidateAll();
-			lastFetchDate = Date.now();
+			lastPingDate = new Date();
 			return newPercent - 100;
 		}
 
@@ -33,7 +33,7 @@
 
 <div
 	class="w-6 h-6 rounded-full overflow-hidden border-2 border-stone-800 transition-transform duration-300"
-	style:transform={makeSmaller ? 'scale(0.75)' : 'scale(1)'}
+	style:transform={currentlyPinging ? 'scale(0.75)' : 'scale(1)'}
 >
 	<div class="w-full h-full" style={`background: conic-gradient(rgb(41, 37, 36) ${percent}%, 0, transparent)`}></div>
 </div>
