@@ -1,9 +1,20 @@
-import { text, integer, sqliteTable } from "drizzle-orm/sqlite-core";
-import { nanoid } from 'nanoid';
+import { z } from 'zod';
 
-export const urls = sqliteTable('urls', {
-    id: text('id').unique().primaryKey().$defaultFn(() => nanoid()),
-    url: text('url').notNull(),
-    timeTaken: integer('timeTaken'),
-    lastFetch: integer('lastFetch').notNull().$defaultFn(() => Date.now()),
+/**
+ * Schema for creating a new URL
+ */
+export const createUrlSchema = z.object({
+    url: z.url('Invalid URL format'),
+    id: z.uuid().default(() => crypto.randomUUID()),
+    createdAt: z.coerce.date().optional()
 });
+
+/**
+ * Schema for deleting a URL by ID
+ */
+export const deleteUrlSchema = z.object({
+    id: z.uuid()
+});
+
+export type CreateUrlInput = z.infer<typeof createUrlSchema>;
+export type DeleteUrlInput = z.infer<typeof deleteUrlSchema>;
