@@ -1,34 +1,18 @@
-import type { Actions } from './$types'
-import { fail } from '@sveltejs/kit'
-import { addUrl, INTERVAL, getLastFetchDate, getUrls, removeUrlByID } from '$lib/services/pinger'
-import { validURL } from '$lib/utils'
+import type { PageServerLoad } from './$types';
+import { prisma } from '$lib/server/prisma';
+import { lastPingAt } from '$lib/server/pinger';
+import { PING_INTERVAL } from '$lib/config';
 
-export async function load() {
+// Pass URLs on load for initial rendering
+export const load: PageServerLoad = async () => {
+    const urls = await prisma.url.findMany({
+        include: { lastPing: true },
+        orderBy: { createdAt: 'asc' }
+    });
+    
     return {
-        interval: INTERVAL,
-        urls: await getUrls(),
-        lastFetchDate: await getLastFetchDate()
+        urls,
+        lastPingDate : lastPingAt,
+        pingInterval: PING_INTERVAL
     };
-}
-
-export const actions: Actions = {
-    add: async ({ request }) => {
-        const data = await request.formData()
-        const url = data.get('url')
-
-        if (typeof url !== 'string' || !validURL(url)) {
-            return fail(400, { url, error: 'Invalid URL specified' });
-        }
-
-        await addUrl(url);
-        return { success: true };
-    },
-
-    delete: async ({ request }) => {
-        const data = await request.formData()
-        const id = data.get('id');
-
-        if (typeof id === 'string') await removeUrlByID(id);
-        return { success: true };
-    }
-}
+};
